Move Parcelable tests for Timestamp into TimestampTest
